Split JitterRng into a separate crate
