+++ rand_jitter/src/lib.rs
@@ -1,28 +1,64 @@
 // Copyright 2018 Developers of the Rand project.
+// Copyright 2013-2015 The Rust Project Developers.
 //
 // Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
 // https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
 // <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
-//
-// Based on jitterentropy-library, http://www.chronox.de/jent.html.
-// Copyright Stephan Mueller <smueller@chronox.de>, 2014 - 2017.
-//
-// With permission from Stephan Mueller to relicense the Rust translation under
-// the MIT license.
 
 //! Non-physical true random number generator based on timing jitter.
+//!
+//! This is a true random number generator, as opposed to pseudo-random
+//! generators. Random numbers generated by `JitterRng` can be seen as fresh
+//! entropy. A consequence is that it is orders of magnitude slower than `OsRng`
+//! and PRNGs (about 10<sup>3</sup>..10<sup>6</sup> slower).
+//!
+//! There are very few situations where using this RNG is appropriate. Only very
+//! few applications require true entropy. A normal PRNG can be statistically
+//! indistinguishable, and a cryptographic PRNG should also be as impossible to
+//! predict.
+//!
+//! Use of `JitterRng` is recommended for initializing cryptographic PRNGs when
+//! `OsRng` is not available.
+//!
+//! `JitterRng` can be used without the standard library, but not conveniently,
+//! you must provide a high-precision timer and carefully have to follow the
+//! instructions of [`new_with_timer`].
+//!
+//! This implementation is based on [Jitterentropy] version 2.1.0.
+//!
+//! Note: There is no accurate timer available on WASM platforms, to help
+//! prevent fingerprinting or timing side-channel attacks. Therefore
+//! [`JitterRng::new()`] always returns an error.
+//!
+//! [`JitterRng::new()`]: struct.JitterRng.html#method.new
+//! [`new_with_timer`]: struct.JitterRng.html#method.new_with_timer
+//! [Jitterentropy]: http://www.chronox.de/jent.html
 
 // Note: the C implementation of `Jitterentropy` relies on being compiled
 // without optimizations. This implementation goes through lengths to make the
 // compiler not optimize out code which does influence timing jitter, but is
 // technically dead code.
 
-use rand_core::{RngCore, CryptoRng, Error, ErrorKind, impls};
+pub extern crate rand_core;
+#[cfg(feature = "log")]
+#[macro_use] extern crate log;
+#[cfg(any(target_os = "macos", target_os = "ios"))]
+extern crate libc;
+#[cfg(target_os = "windows")]
+extern crate winapi;
 
-use core::{fmt, mem, ptr};
-#[cfg(all(feature="std", not(target_arch = "wasm32")))]
+
+#[cfg(not(feature = "log"))]
+#[macro_use] mod dummy_log;
+mod platform;
+mod error;
+
+use rand_core::{RngCore, CryptoRng, Error, impls};
+pub use error::TimerError;
+
+use std::{fmt, mem, ptr};
 use std::sync::atomic::{AtomicUsize, ATOMIC_USIZE_INIT, Ordering};
 
 const MEMORY_BLOCKS: usize = 64;
@@ -31,108 +67,6 @@ const MEMORY_SIZE: usize = MEMORY_BLOCKS * MEMORY_BLOCKSIZE;
 
 /// A true random number generator based on jitter in the CPU execution time,
 /// and jitter in memory access time.
-///
-/// This is a true random number generator, as opposed to pseudo-random
-/// generators. Random numbers generated by `JitterRng` can be seen as fresh
-/// entropy. A consequence is that is orders of magnitude slower than [`OsRng`]
-/// and PRNGs (about 10<sup>3</sup>..10<sup>6</sup> slower).
-///
-/// There are very few situations where using this RNG is appropriate. Only very
-/// few applications require true entropy. A normal PRNG can be statistically
-/// indistinguishable, and a cryptographic PRNG should also be as impossible to
-/// predict.
-///
-/// Use of `JitterRng` is recommended for initializing cryptographic PRNGs when
-/// [`OsRng`] is not available.
-///
-/// `JitterRng` can be used without the standard library, but not conveniently,
-/// you must provide a high-precision timer and carefully have to follow the
-/// instructions of [`new_with_timer`].
-///
-/// This implementation is based on
-/// [Jitterentropy](http://www.chronox.de/jent.html) version 2.1.0.
-///
-/// Note: There is no accurate timer available on Wasm platforms, to help
-/// prevent fingerprinting or timing side-channel attacks. Therefore
-/// [`JitterRng::new()`] is not available on Wasm.
-///
-/// # Quality testing
-///
-/// [`JitterRng::new()`] has build-in, but limited, quality testing, however
-/// before using `JitterRng` on untested hardware, or after changes that could
-/// effect how the code is optimized (such as a new LLVM version), it is
-/// recommend to run the much more stringent
-/// [NIST SP 800-90B Entropy Estimation Suite](
-/// https://github.com/usnistgov/SP800-90B_EntropyAssessment).
-///
-/// Use the following code using [`timer_stats`] to collect the data:
-///
-/// ```no_run
-/// use rand::rngs::JitterRng;
-/// #
-/// # use std::error::Error;
-/// # use std::fs::File;
-/// # use std::io::Write;
-/// #
-/// # fn try_main() -> Result<(), Box<Error>> {
-/// let mut rng = JitterRng::new()?;
-///
-/// // 1_000_000 results are required for the
-/// // NIST SP 800-90B Entropy Estimation Suite
-/// const ROUNDS: usize = 1_000_000;
-/// let mut deltas_variable: Vec<u8> = Vec::with_capacity(ROUNDS);
-/// let mut deltas_minimal: Vec<u8> = Vec::with_capacity(ROUNDS);
-///
-/// for _ in 0..ROUNDS {
-///     deltas_variable.push(rng.timer_stats(true) as u8);
-///     deltas_minimal.push(rng.timer_stats(false) as u8);
-/// }
-///
-/// // Write out after the statistics collection loop, to not disturb the
-/// // test results.
-/// File::create("jitter_rng_var.bin")?.write(&deltas_variable)?;
-/// File::create("jitter_rng_min.bin")?.write(&deltas_minimal)?;
-/// #
-/// # Ok(())
-/// # }
-/// #
-/// # fn main() {
-/// #     try_main().unwrap();
-/// # }
-/// ```
-///
-/// This will produce two files: `jitter_rng_var.bin` and `jitter_rng_min.bin`.
-/// Run the Entropy Estimation Suite in three configurations, as outlined below.
-/// Every run has two steps. One step to produce an estimation, another to
-/// validate the estimation.
-///
-/// 1. Estimate the expected amount of entropy that is at least available with
-///    each round of the entropy collector. This number should be greater than
-///    the amount estimated with `64 / test_timer()`.
-///    ```sh
-///    python noniid_main.py -v jitter_rng_var.bin 8
-///    restart.py -v jitter_rng_var.bin 8 <min-entropy>
-///    ```
-/// 2. Estimate the expected amount of entropy that is available in the last 4
-///    bits of the timer delta after running noice sources. Note that a value of
-///    `3.70` is the minimum estimated entropy for true randomness.
-///    ```sh
-///    python noniid_main.py -v -u 4 jitter_rng_var.bin 4
-///    restart.py -v -u 4 jitter_rng_var.bin 4 <min-entropy>
-///    ```
-/// 3. Estimate the expected amount of entropy that is available to the entropy
-///    collector if both noice sources only run their minimal number of times.
-///    This measures the absolute worst-case, and gives a lower bound for the
-///    available entropy.
-///    ```sh
-///    python noniid_main.py -v -u 4 jitter_rng_min.bin 4
-///    restart.py -v -u 4 jitter_rng_min.bin 4 <min-entropy>
-///    ```
-///
-/// [`OsRng`]: struct.OsRng.html
-/// [`JitterRng::new()`]: struct.JitterRng.html#method.new
-/// [`new_with_timer`]: struct.JitterRng.html#method.new_with_timer
-/// [`timer_stats`]: struct.JitterRng.html#method.timer_stats
 pub struct JitterRng {
     data: u64, // Actual random number
     // Number of rounds to run the entropy collector per 64 bits
@@ -215,62 +149,7 @@ impl Clone for JitterRng {
     }
 }
 
-/// An error that can occur when [`JitterRng::test_timer`] fails.
-///
-/// [`JitterRng::test_timer`]: struct.JitterRng.html#method.test_timer
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum TimerError {
-    /// No timer available.
-    NoTimer,
-    /// Timer too coarse to use as an entropy source.
-    CoarseTimer,
-    /// Timer is not monotonically increasing.
-    NotMonotonic,
-    /// Variations of deltas of time too small.
-    TinyVariantions,
-    /// Too many stuck results (indicating no added entropy).
-    TooManyStuck,
-    #[doc(hidden)]
-    __Nonexhaustive,
-}
-
-impl TimerError {
-    fn description(&self) -> &'static str {
-        match *self {
-            TimerError::NoTimer => "no timer available",
-            TimerError::CoarseTimer => "coarse timer",
-            TimerError::NotMonotonic => "timer not monotonic",
-            TimerError::TinyVariantions => "time delta variations too small",
-            TimerError::TooManyStuck => "too many stuck results",
-            TimerError::__Nonexhaustive => unreachable!(),
-        }
-    }
-}
-
-impl fmt::Display for TimerError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.description())
-    }
-}
-
-#[cfg(feature="std")]
-impl ::std::error::Error for TimerError {
-    fn description(&self) -> &str {
-        self.description()
-    }
-}
-
-impl From<TimerError> for Error {
-    fn from(err: TimerError) -> Error {
-        // Timer check is already quite permissive of failures so we don't
-        // expect false-positive failures, i.e. any error is irrecoverable.
-        Error::with_cause(ErrorKind::Unavailable,
-                              "timer jitter failed basic quality tests", err)
-    }
-}
-
 // Initialise to zero; must be positive
-#[cfg(all(feature="std", not(target_arch = "wasm32")))]
 static JITTER_ROUNDS: AtomicUsize = ATOMIC_USIZE_INIT;
 
 impl JitterRng {
@@ -281,8 +160,10 @@ impl JitterRng {
     /// During initialization CPU execution timing jitter is measured a few
     /// hundred times. If this does not pass basic quality tests, an error is
     /// returned. The test result is cached to make subsequent calls faster.
-    #[cfg(all(feature="std", not(target_arch = "wasm32")))]
     pub fn new() -> Result<JitterRng, TimerError> {
+        if cfg!(target_arch = "wasm32") {
+            return Err(TimerError::NoTimer);
+        }
         let mut state = JitterRng::new_with_timer(platform::get_nstime);
         let mut rounds = JITTER_ROUNDS.load(Ordering::Relaxed) as u8;
         if rounds == 0 {
@@ -314,8 +195,8 @@ impl JitterRng {
     /// # Example
     ///
     /// ```
-    /// # use rand::{Rng, Error};
-    /// use rand::rngs::JitterRng;
+    /// # use rand_jitter::rand_core::{RngCore, Error};
+    /// use rand_jitter::JitterRng;
     ///
     /// # fn try_inner() -> Result<(), Error> {
     /// fn get_nstime() -> u64 {
@@ -332,10 +213,10 @@ impl JitterRng {
     /// let mut rng = JitterRng::new_with_timer(get_nstime);
     /// let rounds = rng.test_timer()?;
     /// rng.set_rounds(rounds); // optional
-    /// let _ = rng.gen::<u64>();
+    /// let _ = rng.next_u64();
     ///
     /// // Ready for use
-    /// let v: u64 = rng.gen();
+    /// let v: u64 = rng.next_u64();
     /// # Ok(())
     /// # }
     ///
@@ -771,45 +652,6 @@ impl JitterRng {
     }
 }
 
-#[cfg(feature="std")]
-mod platform {
-    #[cfg(not(any(target_os = "macos", target_os = "ios",
-                  target_os = "windows",
-                  target_arch = "wasm32")))]
-    pub fn get_nstime() -> u64 {
-        use std::time::{SystemTime, UNIX_EPOCH};
-
-        let dur = SystemTime::now().duration_since(UNIX_EPOCH).unwrap();
-        // The correct way to calculate the current time is
-        // `dur.as_secs() * 1_000_000_000 + dur.subsec_nanos() as u64`
-        // But this is faster, and the difference in terms of entropy is
-        // negligible (log2(10^9) == 29.9).
-        dur.as_secs() << 30 | dur.subsec_nanos() as u64
-    }
-
-    #[cfg(any(target_os = "macos", target_os = "ios"))]
-    pub fn get_nstime() -> u64 {
-        extern crate libc;
-        // On Mac OS and iOS std::time::SystemTime only has 1000ns resolution.
-        // We use `mach_absolute_time` instead. This provides a CPU dependent
-        // unit, to get real nanoseconds the result should by multiplied by
-        // numer/denom from `mach_timebase_info`.
-        // But we are not interested in the exact nanoseconds, just entropy. So
-        // we use the raw result.
-        unsafe { libc::mach_absolute_time() }
-    }
-
-    #[cfg(target_os = "windows")]
-    pub fn get_nstime() -> u64 {
-        extern crate winapi;
-        unsafe {
-            let mut t = super::mem::zeroed();
-            winapi::um::profileapi::QueryPerformanceCounter(&mut t);
-            *t.QuadPart() as u64
-        }
-    }
-}
-
 // A function that is opaque to the optimizer to assist in avoiding dead-code
 // elimination. Taken from `bencher`.
 fn black_box<T>(dummy: T) -> T {
@@ -854,32 +696,3 @@ impl RngCore for JitterRng {
 
 impl CryptoRng for JitterRng {}
 
-#[cfg(test)]
-mod test_jitter_init {
-    use super::JitterRng;
-
-    #[cfg(all(feature="std", not(target_arch = "wasm32")))]
-    #[test]
-    fn test_jitter_init() {
-        use RngCore;
-        // Because this is a debug build, measurements here are not representive
-        // of the final release build.
-        // Don't fail this test if initializing `JitterRng` fails because of a
-        // bad timer (the timer from the standard library may not have enough
-        // accuracy on all platforms).
-        match JitterRng::new() {
-            Ok(ref mut rng) => {
-                // false positives are possible, but extremely unlikely
-                assert!(rng.next_u32() | rng.next_u32() != 0);
-            },
-            Err(_) => {},
-        }
-    }
-
-    #[test]
-    fn test_jitter_bad_timer() {
-        fn bad_timer() -> u64 { 0 }
-        let mut rng = JitterRng::new_with_timer(bad_timer);
-        assert!(rng.test_timer().is_err());
-    }
-}
