+++ src/lib.rs
@@ -69,6 +69,8 @@
 #[macro_use]
 extern crate stdweb;
 
+#[cfg(feature = "rand_jitter")]
+extern crate rand_jitter;
 #[cfg(feature = "rand_os")]
 extern crate rand_os;
