+++ src/rngs/mod.rs
@@ -163,7 +163,6 @@
 pub mod adapter;
 
 #[cfg(feature="std")] mod entropy;
-mod jitter;
 pub mod mock;   // Public so we don't export `StepRng` directly, making it a bit
                 // more clear it is intended for testing.
 mod small;
@@ -171,7 +170,7 @@ mod std;
 #[cfg(feature="std")] pub(crate) mod thread;
 
 
-pub use self::jitter::{JitterRng, TimerError};
+#[cfg(feature="rand_jitter")] pub use rand_jitter::{JitterRng, TimerError};
 #[cfg(feature="std")] pub use self::entropy::EntropyRng;
 
 pub use self::small::SmallRng;
